refactor(fsdp): organize monkey patch modules
Keep version-specific runtime patches in a dedicated package.
