+++ tests/fast-gpu/backends/fsdp_utils/torch_ported/sitecustomize.py
@@ -1,6 +1,6 @@
 import torch
 
-from miles.backends.fsdp_utils.fsdp_param_dtype_patch import apply_param_dtype_map_patch
+from miles.backends.fsdp_utils.monkey_patches.fsdp_param_dtype_patch import apply_param_dtype_map_patch
 
 apply_param_dtype_map_patch()
 torch.use_deterministic_algorithms(True)
